parent access: Make enums scoped
Tiny style improvement that converts enum to scoped enums in order
to prevent implicit conversion bugs.

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3972082
Commit-Queue: Aga Wronska <agawronska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1062900}

+++ chrome/browser/supervised_user/web_approvals_manager.cc
@@ -116,7 +116,8 @@ void WebApprovalsManager::RequestLocalApproval(
           base::TimeTicks::Now()));
 
   if (result != ash::ParentAccessDialogProvider::ShowError::kNone) {
-    LOG(ERROR) << "Error showing ParentAccessDialog: " << result;
+    LOG(ERROR) << "Error showing ParentAccessDialog: "
+               << static_cast<int>(result);
     std::move(callback).Run(false);
     return;
   }
